Added port monitor table
Still need to setup the debugger to break on the Enter on Write/Read
things; but at least now it can update. Try inputting B024 and then
changing the brightness to test. I need to look more into Qt stuff, but
this is probably a lot better if it is done in a TableView?