Fixed stuff.
Still need to fix things for the keypad and interrupts on the port
monitor, since those are parsed differently, but looking better. Oh, and
send the proper signal to tell the main window what port was updated, so
that way it can select it and print a message or something. Looks better
though, and breaks when it is supposed to. Same kind of format will be
used for read/write breakpoints on the main disassembler. I like this
check box thing.

Also, all header information was reorganized and made a lot cleaner, in
addition to putting some things into different directory structures.

It's getting there :)