Added breakpoints to the disassembly view
Now you can set breakpoints from the disasmView. And it shows them
highlighted too. I also was fed up with not being able to jump to
positions, so I added the goto button functionality as well :) Also, you
can right click on the disasmView and set the PC to the current
highlighted (yellow) address. Fun things to play around with, I guess.
:)