Added hex editor panel, fixed bugs, moved files
Also added an icon I may use later for something :P This hex editing
widget has a few bugs; namely: 1) setBaseAddress() does not update the
width of the address column, 2) The cursor does not change with the
horizontal scroll amount. The GUI for it is awful right now; still
deciding how to implement it.