Debugger base things added
These things help get a jump start on issue #15 issues. There's no real
point in sending a pointer to the string for the debugger, as each
character can just be sent to the console by the CPU. It doesn't make
that much of a difference, and other debugger commands can read pointers
because the alternate thread will be halted.