Totally reworked LCD frame drawing... for a tiny color improvement.
Rather than rendering frames in RGB565, frames are now rendered in
RGBA8888. This is because the calculator's screen actually supports
RGB666, so to accurately emulate those least significant red and blue
bits, they need to fit in the encoding. And in testing, I discovered that
the LSB of red and blue should be a copy of the MSB, which is now
implemented.

Performance annoyingly regressed a bit, so please feel free to optimize.
One thing I'm worried about in particular is that, although I chose
RGBA8888 because it's the format the gif writer needs, I'm not sure it's
the format needed to actually display the screen in the emulator. Matching
the frame drawing format to the display format would definitely save
processing time.