Added special menu option for win32 to hide the console window
With newer builds of CEmu, a console window is shown alongside
CEmu for debugging purposes. Some people may not like this, due
to the additional clutter and the fact that if the console window
is closed, CEmu is closed as well.

This commit adds in a bit of a hack - if compiled on Windows (or
rather, for Windows), a special menu option will be added to hide
or show the console window. When selected, it will enable/disable
the console.

Note that when CEmu is spawned from an existing console, the
console will not be hidden. Instead, output will no longer appear
in the console. Re-enabling the console will restore the ability
for future output to be shown.