Attempt to fix MSVC build on AppVeyor
To fix the MSVC build, a new library had to be built to accomodate
the new libpng-apng dependency. All configurations of the library
have been built and uploaded to JFrog/Bintray for use by the
build scripts.

Linking to this new library required some small changes to the
CEmu.pro file - specifically, not using pkg-config in a MSVC
environment, and instead relying on manually inputted LIBS
and INCLUDES to complete the build.

Finally, core/defines.h introduced a new likely()/unlikely() macro
that was incompatible with MSVC. To make things work, this has been
stubbed out for MSVC only.