Add proper qmake defs for MSVC libpng-apng + zlib paths
To make things clearer, from now on, if qmake is invoked with
MSVC generation, qmake will request that LIBPNG_APNG_LIB,
ZLIB_LIB, and LIBPNG_APNG_INCLUDE are defined before allowing
a project file to be generated. These can be specified both
in the environment as well as the qmake commandline, and will
be concatenated to LIBS and INCLUDEPATH respectively.

Finally, finally... fixes #140.