Redo setPortable(bool) and adjust/simplify the rest. (fix #238)
Use absolute paths everywhere, except within the config file when in
portable mode.
Basically we want to guarantee that in memory (those m_* member vars)
all paths are absolute, which simplifies a lot of things.
When reading/writing the config file, then some paths may be relative
to appDir, but that's all.