Fix app getting stuck in QCoreApplication loop on settings migration
When a version change occurs and CEmu detects this, it will popup an
alert on startup prompting the user to migrate or clear their settings.
If they choose to do so, the application will get stuck on the
app.exec().

This is due to the fact that QCoreApplication::exec() is called
without any windows open, so there are no events to trigger it to
exit. Even though the rule is that when the last window is closed,
the QCoreApplication exits, it is still event triggered.

The behavior can be confirmed by breaking on the function
QCoreApplicationPrivate::maybeQuit, which contains the logic necessary
to exit the app when the windows are closed. It is called once for the
dialog box closes, but since there is no event loop running it has no
effect.

Fix is to not even start the event loop unless we pop open a window.