Update MSVC Qt to v5.11.1 from v5.10.1, remove custom libpng-apng
Qt v5.11.1 has been (hopefully!) built successfully as a static
library, so both dynamic and static builds have been upgraded to use
said newer version of Qt.

Minor note - the amount of libraries in Qt has increased from each
release, and v5.11.1 caused the exclusion flags to exceed Windows'
command line limit. Wrote in a "optimizer" to try and replace excluded
libraries with wildcards, and it seems like it works... famous last
words in the face of an inevitable AppVeyor build for this commit!

Another thing - BuildQt511DevArchive.py will likely not be maintained
too much going forward. AppVeyor is pretty good about updating their Qt
versions, and the only reason we've needed this in the past is to build
and deploy pre-release versions of Qt. Maybe it'll be kept around for
now, or maybe even merged into one archive builder...

Finally, we've removed libpng-apng from the downloading and packaging
phases of the AppVeyor deployment since we're using vcpkg's version
now. This should hopefully be a clean breakaway from our custom built
package of libpng-apng.